shift filter pages to components
